Build support for static HTML
- Still missing templating for all images
- Relies on https://github.com/brownplt/code.pyret.org/pull/585
- Would really like to have code.pyret.org as a npm dependency

This uses templating (currently just string replaceAll, should upgrade to
mustache) to fill in an editor.html with the dynamically-generated base URL
provided by VScode.

This gets us one step closer to (a) a sensible build that doesn't require
copying built stuff from CPO (b) a self-contained extension that doesn't rely
on any network connection.

I wanted to push to checkpoint this, there is still an issue that a (!!google)
expression should be !!window.google or similar (it errors on google being
undefined), and maybe some other similar issues. But this is getting close.